Finishing of webscraping and implementation of parallelism
