+++ src/python_files/scanners/alpha.py
@@ -2,51 +2,56 @@ import requests
 from bs4 import BeautifulSoup
 import datetime
 
-location = "Pourhouse"
-log_mode = 1
+log_mode = 0
 log = []
 log_file = "C:/Fall_2021/CSCI_331/Final_Project/Barfly/src/python_files/log_files/"+datetime.datetime.now().strftime("%d_%b_%Y")+".txt"
-output_location = "C:/Fall_2021/CSCI_331/Final_Project/Barfly/src/scraped_files/pourhouse.txt"
+output_location = "C:/Fall_2021/CSCI_331/Final_Project/Barfly/src/scraped_files/alpha.txt"
 
 def time():
     return datetime.datetime.now().strftime("%d/%b/%Y %H:%M:%S")
 
-def pourhouse_food():
 
-    log.append(time() + " Starting Pourhouse Food Function")
+def montana_aleworks_beer():
+
+    location = "Montana Aleworks"
+    
+    #log.append(time() + " Starting Montana Aleworks Beer Function")
 
     output = []
     
-    URL = "https://www.phbozeman.com/"
+    URL = "https://www.montanaaleworks.com/beer"
     page = requests.get(URL)
 
     soup = BeautifulSoup(page.content, "html.parser")
     
     items = soup.find_all("div", "menu-item")
     for item in items:
-        name = "Food"
+        name = "Beer"
         title = item.find("div", "menu-item-title").text.strip()
         description = item.find("div", "menu-item-description")
         if(description == None):
-            log.append(time()+ " ERROR: Description unavailible for " + title)
+            #log.append(time()+ " ERROR: Description unavailible for " + title)
             description = ""
         else:
             description = description.text.strip()
         price = item.find("span", "menu-item-price-top")
         if(price == None):
-            title = title.split("|")
-            price = title[1].strip()
-            title = title[0].strip()
+            price = None
+            #log.append(time()+ " ERROR: Price unavailible for " + title)
         else:
             price = price.text.strip()
+            price = price.split("$")[1:]
+            price = "".join([i.strip() for i in price])
+            price = price.split(" ")
+            price = "".join([i.strip() for i in price])
         if(description != None and price != None):
             output.append("|".join([location, name, title, price, description]))
 
-    log.append(time()+ " Finished Pourhouse Food Function")
+    #log.append(time()+ " Finished Montana Aleworks Beer Function")
     return output
 
 file = open(output_location, "w")
-for item in pourhouse_food():
+for item in montana_aleworks_beer():
     file.write(item + "\n")
 file.close()
 
@@ -56,4 +61,4 @@ if(log_mode):
         file.write(item + "\n")
     file.close()
 
-print("Pourhouse Scan Completed")
+print("Alpha Scan Completed")
