+++ src/scraped_files/bravo.txt
@@ -1,0 +1,50 @@
+Montana Aleworks|Cocktail|not-so-old fashioned|10|fresh thyme, orange and cardamom-infused bourbon, maple syrup, bitters
+Montana Aleworks|Cocktail|ale works marg|9.5|sauza signature blue tequila, gran gala, fresh citrus, tajín spiced rim
+Montana Aleworks|Cocktail|pj marg|10|house-infused pineapple and jalapeño tequila, fresh lime, sweet & sour
+Montana Aleworks|Cocktail|classic mojito|9|cruzan silver rum, muddled mint, simple, fresh lime
+Montana Aleworks|Cocktail|huckleberry mojito|10|vodka, muddled mint, fresh lime, huckleberry puree
+Montana Aleworks|Cocktail|moscow mule|9.5|wildrye distillery vodka
+Montana Aleworks|Cocktail|huckleberry mule|10|vodka, huckleberry puree, ginger beer, fresh lime
+Montana Aleworks|Cocktail|headframe neversweat|6|headframe spirits, butte, mt
+Montana Aleworks|Cocktail|willie’s bighorn|7|willie's distillery, ennis, mt
+Montana Aleworks|Cocktail|glacier fireweed cherry|7|glacier distillery, coram, mt
+Montana Aleworks|Cocktail|grand teton colter’s run|7|grand teton distillery, driggs, id
+Montana Aleworks|Cocktail|wildrye five drops|8|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|whistling andy's|8|whistling andy distillery, big fork, mt
+Montana Aleworks|Cocktail|willie’s canadian|6|willie's distillery, ennis, mt
+Montana Aleworks|Cocktail|montana 1889|6|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|bin 7 wheat|7|dry hills distillery, bozeman, mt
+Montana Aleworks|Cocktail|glacier northfork rye|7|glacier distillery, coram, mt
+Montana Aleworks|Cocktail|highwood rye malt|7|trailhead spirits, billings, mt
+Montana Aleworks|Cocktail|highwood chocolate|7|trailhead spirits, billings, mt
+Montana Aleworks|Cocktail|montgomery sudden wisdom|8|montgomery distillery, missoula, mt
+Montana Aleworks|Cocktail|glacier bear proof huckleberry|7|glacier distillery, coram, mt
+Montana Aleworks|Cocktail|devil's brigade|7|willie's distillery, ennis, mt
+Montana Aleworks|Cocktail|prairie schooner spiced rum|6|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|ramsdell’s parrot rum|8|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|spotted bear agave|8|spotted bear spirits, whitefish, mt
+Montana Aleworks|Cocktail|bozeman arrestado agave|8|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|boss tweed old tom|7|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|anselmo|7|headframe spirits, butte, mt
+Montana Aleworks|Cocktail|whistling andy cucumber|7|whistling andy distillery, big fork, mt
+Montana Aleworks|Cocktail|guardian reserve|7|gulch distillers, helena, mt
+Montana Aleworks|Cocktail|montgomery whyte laydie|7|montgomery distillery, missoula, mt
+Montana Aleworks|Cocktail|ruby river|7|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|great grey owl|7|jackson hole still works, jackson hole, wy
+Montana Aleworks|Cocktail|ginger bear|7|2 fat guys spirits, bozeman, mt
+Montana Aleworks|Cocktail|wildrye|5|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|hollowtop wild raspberry|6|dry hills distillery, bozeman, mt
+Montana Aleworks|Cocktail|wildrye cherry|6|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|montgomery quicksilver|6|montgomery distillery, missoula, mt
+Montana Aleworks|Cocktail|cold springs|6|bozeman spirits distillery, bozeman, mt
+Montana Aleworks|Cocktail|cold springs huckleberry|6|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|cold springs lemon|6|bozeman spirits, bozeman, mt
+Montana Aleworks|Cocktail|grand teton|6|grand teton distillery, driggs, id
+Montana Aleworks|Cocktail|44° north huckleberry|6|44° north distillery, boise, id
+Montana Aleworks|Cocktail|gulch burrone fernet|6|gulch distillery, helena, mt
+Montana Aleworks|Cocktail|willie’s huckleberry cream|6|willie's distillery, ennis, mt
+Montana Aleworks|Cocktail|wildrye apple pie|6|wildrye distillery, bozeman, mt
+Montana Aleworks|Cocktail|willie’s mt honey moonshine|6|willie's distillery, ennis, mt
+Montana Aleworks|Cocktail|headframe orphan girl bourbon cream|6|headframe spirits, butte, mt
+Montana Aleworks|Cocktail|montomery skadi barrel-finished aquavit|7|montgomery distillery, missoula, mt
+Montana Aleworks|Cocktail|willie's coffee cream|6|willie's distillery, ennis, mt
